requirements.txt
torch>=2.0.0
diffsynth
safetensors
Pillow
numpy
modelscope
